+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Master
 * minor improvements
 * 3D plot of streamlines with coloring according to tractometry FA
+* Add pretrained weights for XTRACT tract definitions
+
 
 ## Release 2.1.1